Move event definition and formatting out to a separate file
This is done in preparation for testing, which will in turn be a
preparation for additional modifications.
